Create api and model modules
