+++ .gitignore
@@ -1,2 +1,3 @@
 venv/
 .idea
+__pycache__/
