+++ api/model/Postagem.py
@@ -1,0 +1,27 @@
+from api import flaskApp
+
+db = flaskApp.db
+
+class Postagem(db.Model):
+    __tablename__ = 'postagens'
+    id = db.Column(db.Integer, primary_key=True)
+    titulo = db.Column(db.String(400), nullable=False)
+    texto = db.Column(db.String(400), nullable=False)
+    criador = db.Column(db.Integer, db.ForeignKey('usuarios.id'), nullable=False)
+    categoria = db.Column(db.Integer, db.ForeignKey('categorias.id'), nullable=False)
+    selo = db.Column(db.Boolean, default=False, nullable=False)
+    data = db.Column(db.Time)
+
+    def __init__(self, titulo, texto, criador, categoria):
+        self.titulo = titulo
+        self.texto = texto
+        self.criador = criador
+        self.categoria = categoria
+
+class Categoria(db.Model):
+    __tablename__ = 'categorias'
+    id = db.Column(db.Integer, primary_key=True)
+    nome = db.Column(db.String(80), unique=True, nullable=False)
+
+    def __init__(self, nome):
+        self.nome = nome
